Always Running AI to build model
It runs in the background, capturing your gameplay and measures your accuracy, sensetivity, wrong doings, flicks kinda, how far over you go the target. After the round it will process everything and the model will use this footage to train, also added a video folder to place other clips of other players. Idea is to stort them between good aim and bad aim so the model knows whats good or not and give a rating from 0-100 and how to improve upon your skills in games.